Add default message
When using gloria without any CLI options or commands it doesn't return
any text to the user. I think it would be useful to add a default
message for the user, as it is unclear if something happened or not.
I've added the showHelp function which will print the help if no
commands or options are passed.